+++ packages/apollo-server-core/src/caching.ts
@@ -1,0 +1,4 @@
+export interface PersistedQueryCache {
+  set(key: string, data: string): Promise<any>;
+  get(key: string): Promise<string | null>;
+}