+++ packages/apollo-server-core/src/errors.ts
@@ -184,6 +184,32 @@ export class ForbiddenError extends ApolloError {
   }
 }
 
+export class PersistedQueryNotFoundError extends ApolloError {
+  constructor() {
+    super('PersistedQueryNotFound', 'PERSISTED_QUERY_NOT_FOUND');
+
+    // Set the prototype explicitly.
+    // https://stackoverflow.com/a/41102306
+    Object.setPrototypeOf(this, PersistedQueryNotFoundError.prototype);
+    Object.defineProperty(this, 'name', {
+      value: 'PersistedQueryNotFoundError',
+    });
+  }
+}
+
+export class PersistedQueryNotSupportedError extends ApolloError {
+  constructor() {
+    super('PersistedQueryNotSupported', 'PERSISTED_QUERY_NOT_SUPPORTED');
+
+    // Set the prototype explicitly.
+    // https://stackoverflow.com/a/41102306
+    Object.setPrototypeOf(this, PersistedQueryNotSupportedError.prototype);
+    Object.defineProperty(this, 'name', {
+      value: 'PersistedQueryNotSupportedError',
+    });
+  }
+}
+
 export class BadUserInputError extends ApolloError {
   constructor(message: string, properties?: Record<string, any>) {
     super(message, 'BAD_USER_INPUT', properties);