+++ packages/apollo-server-core/src/types.ts
@@ -6,7 +6,10 @@ import { ListenOptions as HttpListenOptions } from 'net';
 import { GraphQLExtension } from 'graphql-extensions';
 import { EngineReportingOptions } from 'apollo-engine-reporting';
 
-import { GraphQLServerOptions as GraphQLOptions } from './graphqlOptions';
+import {
+  GraphQLServerOptions as GraphQLOptions,
+  PersistedQueryOptions,
+} from './graphqlOptions';
 
 export type Context<T = any> = T;
 export type ContextFunction<T = any> = (
@@ -35,7 +38,6 @@ export interface Config
       | 'validationRules'
       | 'formatResponse'
       | 'fieldResolver'
-      | 'debug'
       | 'cacheControl'
       | 'tracing'
     > {
@@ -48,6 +50,7 @@ export interface Config
   mocks?: boolean | IMocks;
   engine?: boolean | EngineReportingOptions;
   extensions?: Array<() => GraphQLExtension>;
+  persistedQueries?: PersistedQueryOptions | false;
 }
 
 // XXX export these directly from apollo-engine-js